requirements.txt
huggingface_hub
numpy
opencv_python
Pillow
tqdm
